+++ contracts/contracts/protocol/mocks/MockCommitmentGenerator.sol
@@ -8,7 +8,7 @@ contract MockCommitmentGenerator is CommitmentGenerator {
     function internalGenerateCommitment(
         uint256 pubSpendingKeyX,
         uint256 pubSpendingKeyY,
-        uint96 scaledAmount,
+        uint64 scaledAmount,
         uint160 zAssetId,
         uint32 creationTime
     ) external pure returns (bytes32 commitment) {
